selection sort

+++ 2-selection_sort.c
@@ -14,7 +14,7 @@ void selection_sort(int *array, size_t size)
 	size_t i,j;
 	for(i = 0; i < size-1; i++)
 	{
-		int min = i;
+		size_t min = i;
 		for(j = i+1; j < size; j++)
 		{
 			if(array[j] < array[min])
